author-manager/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow_sqlalchemy
psycopg2
python-dotenv
marshmallow
passlib
flask-jwt-extended
itsdangerous
Flask-Mail